tree: move auth stuff
